requirements.txt
#Django
django==1.11.6

#Django Rest Framework
djangorestframework==3.7.0
djangorestframework_jwt==1.11.0
django_cors_headers==2.1.0

#MySQL
mysqlclient==1.3.12